requirements.txt
Django==4.2.7
mongoengine==0.29.3
pymongo==4.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
Pillow==11.2.1
opencv-python==4.10.0.84
pytesseract==0.3.13
pypdf==5.4.0
cloudinary==1.41.0
python-dotenv==1.0.1
